+++ packages/@tailwindcss-browser/vitest.config.ts
@@ -1,0 +1,7 @@
+import { defineConfig } from 'vitest/config'
+
+export default defineConfig({
+  test: {
+    exclude: ['**/*.spec.?(c|m)[jt]s?(x)', '**/node_modules/**'],
+  },
+})